+++ lib/src/services/core/message_reaction.dart
@@ -4,25 +4,25 @@ extension ZIMKitCoreMessageReaction on ZIMKitCore {
   Future<void> addMessageReaction(
       ZIMKitMessage message, String reactionType) async {
     if (reactionType.isEmpty) {
-      ZIMKitLogger.warning("addMessageReaction: reactionType is empty");
+      ZIMKitLogger.logWarn("addMessageReaction: reactionType is empty");
       return Future.error(
           PlatformException(code: '-1', message: 'reactionType is empty'));
     }
-    ZIMKitLogger.info(
+    ZIMKitLogger.logInfo(
         'addMessageReaction: $reactionType, messageID:${message.info.messageID}');
 
     await ZIM
         .getInstance()!
         .addMessageReaction(reactionType, message.zim)
         .then((result) {
-      ZIMKitLogger.info(
+      ZIMKitLogger.logInfo(
           'addMessageReaction: success, $reactionType, messageID:${message.info.messageID}');
     }).catchError((error) {
       return checkNeedReloginOrNot(error).then((retryCode) {
         if (retryCode == 0) {
-          ZIMKitLogger.info('relogin success, retry addMessageReaction');
+          ZIMKitLogger.logInfo('relogin success, retry addMessageReaction');
         } else {
-          ZIMKitLogger.severe(
+          ZIMKitLogger.logError(
               'addMessageReaction: failed, error:$error, $reactionType, messageID:${message.info.messageID}');
           throw error;
         }
@@ -33,25 +33,25 @@ extension ZIMKitCoreMessageReaction on ZIMKitCore {
   Future<void> deleteMessageReaction(
       ZIMKitMessage message, String reactionType) async {
     if (reactionType.isEmpty) {
-      ZIMKitLogger.warning("deleteMessageReaction: reactionType is empty");
+      ZIMKitLogger.logWarn("deleteMessageReaction: reactionType is empty");
       return Future.error(
           PlatformException(code: '-1', message: 'reactionType is empty'));
     }
-    ZIMKitLogger.info(
+    ZIMKitLogger.logInfo(
         'deleteMessageReaction: $reactionType, messageID:${message.info.messageID}');
 
     await ZIM
         .getInstance()!
         .deleteMessageReaction(reactionType, message.zim)
         .then((result) {
-      ZIMKitLogger.info(
+      ZIMKitLogger.logInfo(
           'deleteMessageReaction: success, $reactionType, messageID:${message.info.messageID}');
     }).catchError((error) {
       return checkNeedReloginOrNot(error).then((retryCode) {
         if (retryCode == 0) {
-          ZIMKitLogger.info('relogin success, retry deleteMessageReaction');
+          ZIMKitLogger.logInfo('relogin success, retry deleteMessageReaction');
         } else {
-          ZIMKitLogger.severe(
+          ZIMKitLogger.logError(
               'deleteMessageReaction: failed, error:$error, $reactionType, messageID:${message.info.messageID}');
           throw error;
         }
@@ -61,7 +61,7 @@ extension ZIMKitCoreMessageReaction on ZIMKitCore {
 
   void onMessageReactionsChanged(
       ZIM zim, List<ZIMMessageReaction> reactions) async {
-    ZIMKitLogger.info('onMessageReactionsChanged: ${reactions.length}');
+    ZIMKitLogger.logInfo('onMessageReactionsChanged: ${reactions.length}');
     if (reactions.isEmpty) return;
 
     for (final reaction in reactions) {
@@ -71,7 +71,7 @@ extension ZIMKitCoreMessageReaction on ZIMKitCore {
       final messageList = db.messages(conversationID, conversationType);
 
       if (messageList.notInited) {
-        ZIMKitLogger.info(
+        ZIMKitLogger.logInfo(
             'onMessageReactionsChanged: notInited, loadMessageList first');
         await getMessageListNotifier(conversationID, conversationType);
       }