+++ lib/zego_zimkit.dart
@@ -1,9 +1,5 @@
 library zego_zimkit;
 
-/// @todo@yoer don't export third library
-export 'package:file_picker/file_picker.dart';
-export 'package:zego_zim/zego_zim.dart';
-
 export 'src/callkit/defines.dart';
 export 'src/components/components.dart';
 export 'src/pages/pages.dart';